Support for multiple web applications
Remi is a library that strictly only supports one-page web servers. Working around that would require to pass BaseRequestHandler.request (i.e., HTTP socket) to instance of the correct class derived from remi.App (a http.server.StreamRequestHandler based on BaseRequestHandler) based on relative URLs (self.path) AND:

1. to overwrite do_*() methods in such a way that they accept non-root relative paths like /say or /admin and
2. to rewrite all methods of remi.App accessing existing client and instance global variables in su ch a way that they allow multiple App instances to be defined for one client/session.

That could only hardly be done and would even make the project dependent on a fixed version of remi and possibly Python too. Very bad idea.

Thus, a proper reverse proxy server with process-separated remi applications with their own client, sesion and instance accounting was used. For that, Python package mitmproxy reverse proxy was used. It has an awflu lot of dependencies, though, and is quite heavyweight.

Arguably, an OS-level reverse proxy like nginx would be better suited for the job. That way, however, collision with a potentially existing instance of nginx within the system would have to be prevented, the configuration would have to be generated on the fly from the Python project and, most importantly, its proper life cycle would have to be controlled by the Python project as well, even if it crashed.